feat: Add group ID column to analysis data
The code changes in this commit add a new column, `group_id`, to the analysis data. This column is used to identify groups of data points based on changes in the `volatile_tag` column. The `update_group_id` function is introduced to update the `group_id` based on the defined conditions. This enhancement improves the analysis capabilities by providing additional insights into the data grouping.