Refactor streamlit.py to update period and std_period options
This commit updates the streamlit.py file in the volatile_analysis directory to modify the options for the "Period" and "STDV Period" multiselect inputs. The previous options were [1764, 1008] for "Period" and [5, 10, 20, 40, 80] for "STDV Period". The new options are [5, 10, 20, 40, 80] for "Period" and [1764, 1008] for "STDV Period". This change aligns the options with the desired values and improves the usability of the application.