Refactor code to update TIMEFRAME_OPTIONS and STRATEGY_DB_PATH
This commit updates the `.env` file to include additional options for the `TIMEFRAME_OPTIONS` variable. The new options are `2,3,30,5,7`. Additionally, the `STRATEGY_DB_PATH` variable has been updated to `C:\Users\Calathea Finserv\Desktop\Abhishek\20240814` in the `.env` file. These changes improve the flexibility and customization options for the application.